+++ src/datepicker-next/datepicker-next.directive.ts
@@ -1,22 +1,7 @@
-import {
-    Directive,
-    OnInit,
-    ElementRef,
-    Renderer2,
-    ViewContainerRef,
-    Input,
-    ComponentRef,
-    Output,
-    EventEmitter,
-    forwardRef,
-    OnChanges,
-    AfterContentInit
-} from '@angular/core';
-import { ComponentLoaderFactory, ComponentLoader } from 'ngx-bootstrap/component-loader';
-import { DatePipe, registerLocaleData } from '@angular/common';
-import localeZhHans from '@angular/common/locales/zh-Hans';
+import { Directive, OnInit, ElementRef, Renderer2, ViewContainerRef, Input, forwardRef, AfterContentInit } from '@angular/core';
+import { ComponentLoaderFactory } from 'ngx-bootstrap/component-loader';
+import { DatePipe } from '@angular/common';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { isObject, isNumber, isDate, inputValueToBoolean } from '../util/helpers';
 import { ThyPositioningService, PlacementTypes } from '../positioning/positioning.service';
 import { DatepickerNextValueInfo } from './datepicker-next.interface';
 
@@ -27,7 +12,7 @@ const FORMAT_RULES = {
 };
 
 @Directive({
-    selector: '[thyDatepickerNext]',
+    selector: '[thyDatepickerNextAction]',
     providers: [
         {
             provide: NG_VALUE_ACCESSOR,