+++ src/shared/shared.module.ts
@@ -1,17 +1,17 @@
 import { Component, NgModule } from '@angular/core';
 import { ThyTranscludeDirective } from './ng-transclude.directive';
 import { ThyTranslate } from './translate';
-import { ThyDraggableDirective } from './draggable.directive';
+// import { ThyDraggableDirective } from './draggable.directive';
 import { ThyPositioningService } from '../positioning/positioning.service';
 
 @NgModule({
     declarations: [
         ThyTranscludeDirective,
-        ThyDraggableDirective
+        // ThyDraggableDirective
     ],
     exports: [
         ThyTranscludeDirective,
-        ThyDraggableDirective
+        // ThyDraggableDirective
     ],
     providers: [
         ThyTranslate,