+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 7.0.7
+`thyDatepicker` 时间范围选择样式的修改
+
+## 7.0.6
+`thyGrid` 新增支持 `thyRowClassName`
+`table` 样式支持排除某一行拖拽样式 `table-draggable-ignore-item`
+
+## 7.0.5
+`thy-date-range` 当双向绑定的值为空时，根据 `dateRanges` 设置日期区间；当双向绑定的值不为空时，不修改双向绑定的值。
+
+## 7.0.4
+修复发布错误。
+
+## 7.0.2
+`thyDatepicker` 新增支持设置最大值 `thyMaxDate`、最小值 `thyMaxDate`。
+
+## 7.0.1
+注销：ThyDraggableDirective，sky使用报错
+
 ## 7.0.0
 依赖库升级到 Angular 7.2.6， bootstrap 4.3.1。
 主版本号，更改为随 Angular 主版本。